requirements.txt
aiohttp~=3.9.3
asyncio~=3.4.3
click~=8.1.7
PyYAML~=6.0.1
pydantic~=2.6.1
mlflow~=2.10.2
smart-open~=6.4.0
sentence-transformers~=2.5.1
# torch==2.1.0+cpu
tqdm~=4.66.2
requests~=2.31.0
boto3~=1.34.60
openai>=1.0
together==1.1.3
